requirements.txt
asgiref==3.5.0
autopep8==1.6.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Django==3.2
django-ckeditor==6.3.2
django-cleanup==6.0.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-jalali==6.0.0
django-js-asset==2.0.0
django-silk==4.3.0
django-widget-tweaks==1.4.12
flake8==4.0.1
gprof2dot==2021.2.21
idna==3.3
isort==5.10.1
jdatetime==4.1.0
Jinja2==3.1.1
kavenegar==1.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
psycopg2==2.9.3
psycopg2-binary==2.8.6
gunicorn==20.0.4

